Conflict resolution keeps master's Eth62 transaction rescheduling and
Eth65 CreatePooledTransactionsMessage, dropping the unrelated
Task -> ValueTask change.

Review follow-ups:
- Key stats by the wrapped message type, not the wrapper's generic type.
  Type.Name of a closed generic is "SyncServeTaskRequest`2" for every
  instantiation, so the drop log and stats dump were ambiguous and lost
  the information the removed `source` argument carried.
- Replace the ConcurrentDictionary counters with a flat int[] indexed by
  the dense task id. Updates were taking a bucket Monitor lock on both
  the enqueue and the dequeue path; they are now a single interlocked add
  (~14x faster uncontended, never slower under contention).
- Count before the channel write so a consumer cannot decrement ahead of
  the increment, and undo the count when the write fails.
- Fixed-capacity type registry: drops the lock, the growth loop and the
  race where a concurrent grow could drop a just-registered type.
- Drop the `in` modifier added to IBackgroundTaskScheduler.TryScheduleTask:
  it blocks implementations from capturing the request in a lambda and the
  request is copied into Activity.Rent regardless.
- Make both stats tests deterministic by parking the single worker instead
  of racing Task.Delay; removes the [Retry(3)].
- Trim the XML docs and drop the s_ prefixes.

+++ src/Nethermind/Nethermind.AuRa.Test/AuRaStepCalculatorTests.cs
@@ -20,7 +20,7 @@ namespace Nethermind.AuRa.Test
         {
             ManualTimestamper manualTimestamper = new(DateTime.UtcNow);
             AuRaStepCalculator calculator = new(GetStepDurationsForSingleStep(stepDuration), manualTimestamper, LimboLogs.Instance);
-            long step = calculator.CurrentStep;
+            ulong step = calculator.CurrentStep;
             manualTimestamper.Add(calculator.TimeToNextStep);
             Assert.That(calculator.CurrentStep, Is.EqualTo(step + 1), calculator.TimeToNextStep.ToString());
         }
@@ -41,21 +41,21 @@ namespace Nethermind.AuRa.Test
             DateTimeOffset time = DateTimeOffset.FromUnixTimeMilliseconds(milliSeconds);
             ManualTimestamper timestamper = new(time.UtcDateTime);
             AuRaStepCalculator calculator = new(GetStepDurationsForSingleStep(stepDuration), timestamper, LimboLogs.Instance);
-            Assert.That(calculator.CurrentStep, Is.EqualTo(time.ToUnixTimeSeconds() / stepDuration));
+            Assert.That(calculator.CurrentStep, Is.EqualTo((ulong)(time.ToUnixTimeSeconds() / stepDuration)));
         }
 
-        [TestCase(100000060005L, 2, 50000030)]
-        [TestCase(100000060005L, 2, 50000000)]
-        [TestCase(100000060005L, 2, 50000031)]
-        [TestCase(100000060005L, 2, 50000035)]
-        public void time_to_step_is_calculated_correctly(long milliSeconds, int stepDuration, int checkedStep)
+        [TestCase(100000060005L, 2, 50000030UL)]
+        [TestCase(100000060005L, 2, 50000000UL)]
+        [TestCase(100000060005L, 2, 50000031UL)]
+        [TestCase(100000060005L, 2, 50000035UL)]
+        public void time_to_step_is_calculated_correctly(long milliSeconds, int stepDuration, ulong checkedStep)
         {
-            const long currentStep = 50000030;
+            const ulong currentStep = 50000030UL;
             TimeSpan timeToNextStep = TimeSpan.FromMilliseconds(1995);
             DateTimeOffset time = DateTimeOffset.FromUnixTimeMilliseconds(milliSeconds);
             ManualTimestamper timestamper = new(time.UtcDateTime);
             AuRaStepCalculator calculator = new(GetStepDurationsForSingleStep(stepDuration), timestamper, LimboLogs.Instance);
-            TimeSpan expected = checkedStep <= currentStep ? TimeSpan.FromMilliseconds(0) : TimeSpan.FromSeconds((checkedStep - currentStep - 1) * stepDuration) + timeToNextStep;
+            TimeSpan expected = checkedStep <= currentStep ? TimeSpan.FromMilliseconds(0) : TimeSpan.FromSeconds((double)(checkedStep - currentStep - 1) * stepDuration) + timeToNextStep;
             TestContext.Out.WriteLine($"Expected time to step {checkedStep} is {expected}");
             Assert.That(calculator.TimeToStep(checkedStep), Is.EqualTo(expected));
         }
@@ -83,8 +83,8 @@ namespace Nethermind.AuRa.Test
         {
             DateTimeOffset now = DateTimeOffset.FromUnixTimeSeconds(BaseOffset);
             ManualTimestamper timestamper = new(now.UtcDateTime);
-            Dictionary<long, long> stepDurations = secondsDurations.ToDictionary(
-                kvp => kvp.SecondsOffset == 0 ? 0 : kvp.SecondsOffset + now.ToUnixTimeSeconds(),
+            Dictionary<ulong, long> stepDurations = secondsDurations.ToDictionary(
+                kvp => kvp.SecondsOffset == 0 ? 0UL : (ulong)(kvp.SecondsOffset + now.ToUnixTimeSeconds()),
                 kvp => kvp.StepDuration);
             AuRaStepCalculator calculator = new(stepDurations, timestamper, LimboLogs.Instance);
             timestamper.Add(TimeSpan.FromSeconds(second));
@@ -92,6 +92,7 @@ namespace Nethermind.AuRa.Test
             Assert.That(calculator.TimeToNextStep, Is.EqualTo(TimeSpan.FromSeconds(expectedSecondsToNextStep)));
         }
 
-        private IDictionary<long, long> GetStepDurationsForSingleStep(long stepDuration) => new Dictionary<long, long>() { { 0, stepDuration } };
+        private IDictionary<ulong, long> GetStepDurationsForSingleStep(long stepDuration) =>
+            new Dictionary<ulong, long>() { { 0UL, stepDuration } };
     }
 }
